Files to open a connection to a proximity service within the space and forward the appropriate messages to it (location updates, proximity query requests) and respond to messages delivered from it (proximity updates get delivered to the space and then reforwarded

+++ libcore/src/util/UUID.cpp
@@ -74,7 +74,7 @@ std::string UUID::rawHexData()const{
     }
     return retval;
 }
-unsigned int UUID::hash() const {
+size_t UUID::hash() const {
     uint64 a;
     uint64 b;
     memcpy(&a,mData.begin(),sizeof(a));
